working for points in loop?, fails sometimes for points out of loop?

+++ include/trainserver.h
@@ -21,14 +21,14 @@ typedef struct TrainServerMsgSimple{
     train_arg_type type;
     uint32_t arg1;
     uint32_t arg2;
-    uint8_t arg3;
+    uint32_t arg3;
 } TrainServerMsgSimple;
 
 typedef struct TrainServerMsg{
     train_arg_type type;
     uint32_t arg1;
     uint32_t arg2;
-    uint8_t arg3;
+    uint32_t arg3;
     char data[300];
 } TrainServerMsg;
